Add stages/generate and stages/chat
# Conflicts:
#	sgl-router/src/routers/grpc/stages/preparation.rs
#	sgl-router/src/routers/grpc/stages/request_building.rs
#	sgl-router/src/routers/grpc/stages/response_processing.rs
