+++ sgl-router/src/routers/grpc/stages/generate/request_building.rs
@@ -1,0 +1,78 @@
+//! Generate request building stage: Build proto GenerateRequest for generate requests
+
+use async_trait::async_trait;
+use axum::response::Response;
+use uuid::Uuid;
+
+use crate::routers::grpc::{
+    context::{ClientSelection, RequestContext, WorkerSelection},
+    stages::PipelineStage,
+    utils,
+};
+
+/// Generate request building stage
+///
+/// Extracts generate-specific request building logic from the old unified RequestBuildingStage.
+pub struct GenerateRequestBuildingStage {
+    inject_pd_metadata: bool,
+}
+
+impl GenerateRequestBuildingStage {
+    pub fn new(inject_pd_metadata: bool) -> Self {
+        Self { inject_pd_metadata }
+    }
+}
+
+#[async_trait]
+impl PipelineStage for GenerateRequestBuildingStage {
+    async fn execute(&self, ctx: &mut RequestContext) -> Result<Option<Response>, Response> {
+        let prep = ctx
+            .state
+            .preparation
+            .as_ref()
+            .ok_or_else(|| utils::internal_error_static("Preparation not completed"))?;
+
+        let clients = ctx
+            .state
+            .clients
+            .as_ref()
+            .ok_or_else(|| utils::internal_error_static("Client acquisition not completed"))?;
+
+        let generate_request = ctx.generate_request_arc();
+
+        // Get client for building request (use prefill client if PD mode)
+        let builder_client = match clients {
+            ClientSelection::Single { client } => client,
+            ClientSelection::Dual { prefill, .. } => prefill,
+        };
+
+        // Build generate request
+        let request_id = generate_request
+            .rid
+            .clone()
+            .unwrap_or_else(|| format!("gen-{}", Uuid::new_v4()));
+
+        let mut proto_request = builder_client
+            .build_plain_generate_request(
+                request_id,
+                &generate_request,
+                prep.original_text.clone(),
+                prep.token_ids.clone(),
+            )
+            .map_err(utils::bad_request_error)?;
+
+        // Inject PD metadata if needed
+        if self.inject_pd_metadata {
+            if let WorkerSelection::Dual { prefill, .. } = ctx.state.workers.as_ref().unwrap() {
+                super::super::common::helpers::inject_bootstrap_metadata(&mut proto_request, prefill);
+            }
+        }
+
+        ctx.state.proto_request = Some(proto_request);
+        Ok(None)
+    }
+
+    fn name(&self) -> &'static str {
+        "GenerateRequestBuilding"
+    }
+}
